+++ RELEASE_NOTES.md
@@ -4,6 +4,10 @@
 
 * `Dispatch.end_time` has been added to the `Dispatch` class, which is the time when the dispatch ended as calculated by the server. `dispatch-cli` will also print this time.
 
+## Upgrading
+
+* Renamed `Client` class to `DispatchApiClient` for clarity.
+
 ## Bug Fixes
 
 * Fix that `dispatch-cli stream` would try to print an event as dispatch, causing an exception.